Rename ArtistSearchView to just SearchView and Group fiels together
